### 2026-01-13 #9762 @editorialbot
[PRE REVIEW]: MF6RTM: a python package for predictive reactive transport modeling via the MODFLOW 6 and PHREEQC APIs
**Submitting author:** @p-ortega (<a href="http://orcid.org/0000-0003-4091-4221">Pablo Ortega-Tong</a>)
**Repository:** https://github.com/p-ortega/mf6rtm
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.5.0
