+++ JOSS/paper.bib
@@ -29,7 +29,7 @@ abstract = {An open-source, scalable and model-independent (non-intrusive) imple
 @article{Parkhurst2015,
    abstract = {PhreeqcRM is a geochemical reaction module designed specifically to perform equilibrium and kinetic reaction calculations for reactive transport simulators that use an operator-splitting approach. The basic function of the reaction module is to take component concentrations from the model cells of the transport simulator, run geochemical reactions, and return updated component concentrations to the transport simulator. If multicomponent diffusion is modeled (e.g., Nernst–Planck equation), then aqueous species concentrations can be used instead of component concentrations. The reaction capabilities are a complete implementation of the reaction capabilities of PHREEQC. In each cell, the reaction module maintains the composition of all of the reactants, which may include minerals, exchangers, surface complexers, gas phases, solid solutions, and user-defined kinetic reactants. PhreeqcRM assigns initial and boundary conditions for model cells based on standard PHREEQC input definitions (files or strings) of chemical compositions of solutions and reactants. Additional PhreeqcRM capabilities include methods to eliminate reaction calculations for inactive parts of a model domain, transfer concentrations and other model properties, and retrieve selected results. The module demonstrates good scalability for parallel processing by using multiprocessing with MPI (message passing interface) on distributed memory systems, and limited scalability using multithreading with OpenMP on shared memory systems. PhreeqcRM is written in C++, but interfaces allow methods to be called from C or Fortran. By using the PhreeqcRM reaction module, an existing multicomponent transport simulator can be extended to simulate a wide range of geochemical reactions. Results of the implementation of PhreeqcRM as the reaction engine for transport simulators PHAST and FEFLOW are shown by using an analytical solution and the reactive transport benchmark of MoMaS.},
    author = {David L Parkhurst and Laurin Wissmeier},
-   doi = {https://doi.org/10.1016/j.advwatres.2015.06.001},
+   doi = {10.1016/j.advwatres.2015.06.001},
    issn = {0309-1708},
    journal = {Advances in Water Resources},
    pages = {176-189},
@@ -203,7 +203,7 @@ abstract = {An open-source, scalable and model-independent (non-intrusive) imple
     volume = {5},
     issue = {51},
     pages = {2317},
-    doi = {https://doi.org/10.21105/joss.02317},
+    doi = {10.21105/joss.02317},
     year = {2020}
 }
 @article{Bakker2023,
@@ -213,6 +213,6 @@ abstract = {An open-source, scalable and model-independent (non-intrusive) imple
     volume = {54},
     issue = {5},
     pages = {733-739},
-    doi = {https://doi.org/10.1111/gwat.12413},
+    doi = {10.1111/gwat.12413},
     year = {2023}
 }
