### 2026-01-28 4ea250c
rewrite for new JOSS scope rules
