+++ JOSS/paper.md
@@ -43,13 +43,13 @@ affiliations:
    index: 5
  - name: Coastal and Hydraulics Lab, Engineer Research and Developement Center, Vicksburg, MS, USA
    index: 6
-date: 25 December 2025
+date: 28 January 2026
 bibliography: paper.bib
 ---
 
 # Summary
 
-Reactive transport modeling plays a central role in characterizing and predicting the coupled behavior of groundwater flow, solute transport, and geochemical reactions in subsurface systems [@Prommer2019]. This paper presents MF6RTM (MODFLOW 6 Reactive Transport Module), a Python package that tightly couples MODFLOW 6 [@Langevin2024], the current generation of the MODFLOW groundwater flow and transport code family, with PHREEQC [@Appelo2010], a widely used geochemical modeling engine. The coupling is achieved through the MODFLOWAPI [@Hughes2022] and PHREEQCRM [@Parkhurst2015] APIs, which use the Basic Model Interface (BMI) version 2.0 [@Hutton2020] to enable efficient and consistent data exchange between hydraulic, transport, and geochemical components during simulation.
+Reactive transport modeling (RTM) plays a central role in characterizing and predicting the coupled behavior of groundwater flow, solute transport, and geochemical reactions in subsurface systems [@Prommer2019]. This paper presents MF6RTM (MODFLOW 6 Reactive Transport Module), a Python package that tightly couples MODFLOW 6 [@Langevin2024], the current generation of the MODFLOW groundwater flow and transport code family, with PHREEQC [@Appelo2010], a widely used geochemical modeling engine. The coupling is achieved through the MODFLOWAPI [@Hughes2022] and PHREEQCRM [@Parkhurst2015] APIs, which use the Basic Model Interface (BMI) version 2.0 [@Hutton2020] to enable efficient and consistent data exchange between hydraulic, transport, and geochemical components during simulation.
 
 The software provides a unified computational environment for simulating a wide range of reactive transport processes, including contaminant migration, mineral dissolution and precipitation, and redox transformations. It also supports the core features of both MODFLOW 6 and PHREEQC modeling software packages, allowing users to represent complex hydrogeological conditions and geochemical systems.
 
@@ -57,25 +57,25 @@ In addition, MF6RTM includes an improved input and output system that can extern
 
 Together, these features make MF6RTM a versatile and robust framework for predictive reactive transport modeling in hydrogeological and environmental applications.
 
+# State of the Field
 
-# Statement of Need
+Several tools for reactive transport modeling exist that have coupled groundwater flow and solute transport simulators with geochemical engines or implemented fully integrated reactive transport solutions. A few actively developed open-source and standalone (implicitly coupled) reactive transport software systems are noteworthy, including CrunchFlow [@Steefel2014], PFLOTRAN [@Hammond2022], and OpenGeoSys [@Kolditz2012]. Other open-source software has explicitly coupled transport and reaction models, including PHAST [@Parkhurst2010], PHT3D [@Prommer2003], and eSTOMP [@Nieplocha2006]. For a more comprehensive overview, see the review by Steefel [@Steefel2015].
 
-Several tools have coupled flow-and-transport simulators with geochemical engines or implemented fully integrated reactive transport solutions. A few actively developed open-source and standalone (implicitly coupled) reactive transport software systems are noteworthy, including CrunchFlow [@Steefel2014], PFLOTRAN [@Hammond2022], and OpenGeoSys [@Kolditz2012]. Other open-source software has explicitly coupled transport and reaction models, including PHAST [@Parkhurst2010], PHT3D [@Prommer2003], and eSTOMP [@Nieplocha2006]. For a more comprehensive overview, see the review by Steefel [@Steefel2015].
[…]
