### 2026-01-28 7c6909e
polish ai disclosure
