### 2026-01-28 96bccb3
add refs to uncertainty statement
