+++ JOSS/paper.bib
@@ -216,3 +216,30 @@ abstract = {An open-source, scalable and model-independent (non-intrusive) imple
     doi = {10.1111/gwat.12413},
     year = {2023}
 }
+@article{Langevin2012,
+   author = {Christian D. Langevin and Sorab Panday},
+   doi = {10.1111/j.1745-6584.2012.00937.x},
+   issn = {0017-467X},
+   issue = {3},
+   journal = {Groundwater},
+   month = {5},
+   pages = {334-339},
+   title = {Future of Groundwater Modeling},
+   volume = {50},
+   url = {https://ngwa.onlinelibrary.wiley.com/doi/10.1111/j.1745-6584.2012.00937.x},
+   year = {2012}
+}
+@article{White2017,
+   abstract = { Article impact statement : A simple reordering of the modeling workflow, putting the forecast to the forefront, will improve modeling efficiency and effectiveness. },
+   author = {Jeremy T. White},
+   doi = {10.1111/gwat.12558},
+   issn = {0017-467X},
+   issue = {5},
+   journal = {Groundwater},
+   month = {9},
+   pages = {660-664},
+   title = {Forecast First: An Argument for Groundwater Modeling in Reverse},
+   volume = {55},
+   url = {https://ngwa.onlinelibrary.wiley.com/doi/10.1111/gwat.12558},
+   year = {2017}
+}
