+++ JOSS/paper.md
@@ -65,7 +65,7 @@ Despite this comprehensive ecosystem, no open-source software has ever coupled t
 
 Previous PHREEQC couplings within the MODFLOW ecosystem include PHT3D for MODFLOW-2005 [@Prommer2003] and PHT-USG for MODFLOW-USG [@Panday2013]. PHT3D has seen extensive use in both academia and practice [@Appelo2010], while PHT-USG has gained traction more recently, particularly between practioners working with MODFLOW-USG. A key limitation of both approaches is that they require modification of the underlying source code to enable the coupling. This imposes a heavy maintenance burden and has effectively freezed these coupled systems to older software versions. Indeed, both PHT3D and PHT-USG still rely on PHREEQC-2 and updating to the latest PHREEQC version 3 [@Parkhurst2013] through the PHREEQCRM library would require substantial refactoring. As MODFLOW 6 and PHREEQC continue to expand in capability and adoption, there is a clear need for a modern, open-source coupling that preserves transparency, extensibility, and computational efficiency. MF6RTM addresses this need by providing a fully open, API-based integration between MODFLOW 6 and PHREEQC. This design eliminates custom file-based workflows, reduces opportunities for error, and enables users to construct complex reactive transport simulations directly in Python. 
 
-Moreover, there is a growing expectation that groundwater models, both reactive and non-reactive, explicitly represent uncertainty and support automated history-matching and optimization. Historically, most reactive transport workflows have relied on manual or ad hoc modification of input files to perform sensitivity analyses or history-matching, creating a substantial burden for modelers and limiting reproducibility. By enabling array-style input and output files, MF6RTM streamlines compatibility with uncertainty and optimization tools, and supports rigorous uncertainty analysis and multi-objective optimization. MF6RTM therefore fills an important gap in the hydrogeologic modeling ecosystem, providing researchers and practitioners with an accessible, reliable, and high-performance MODFLOW-based framework for fully scripted reactive transport modeling.
+Moreover, there is a growing expectation that groundwater models, both reactive and non-reactive, explicitly represent uncertainty and support automated history-matching and optimization [@Langevin2012,@White2017]. Historically, most reactive transport workflows have relied on manual or ad hoc modification of input files to perform sensitivity analyses or history-matching, creating a substantial burden for modelers and limiting reproducibility. By enabling array-style input and output files, MF6RTM streamlines compatibility with uncertainty and optimization tools, and supports rigorous uncertainty analysis and multi-objective optimization. MF6RTM therefore fills an important gap in the hydrogeologic modeling ecosystem, providing researchers and practitioners with an accessible, reliable, and high-performance MODFLOW-based framework for fully scripted reactive transport modeling.
 
 
 # Software Design
