### 2026-01-28 7faa97e
fix citation attempt 2
