### 2026-01-28 f16b30b
split into need and state of field
