### 2026-01-28 #9762 @p-ortega
minor edits in main manuscript
