### 2026-03-17 #9762 @kthyng
- [ ]  [Not a minor utility](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance)

New submission criteria since journal scope changes in January 2026 (more details in [submission docs](https://joss.readthedocs.io/en/latest/submitting.html#pre-review-screening-criteria)):
- [ ] Public development history ≥ 6 months
- [ ] Demonstrated research impact
- [ ] Open source practices
- [ ] Iterative development
- [ ] Check for borderline cases
