### 2026-03-17 #9762 @kthyng
@p-ortega

* Do you have tests for your code? Please point out the directory.
* What does the following mean? Do you have other code bases within your code base? and if so, why not install them as a dependency?

> All dependencies and executables are included in this repo. This package extensively uses [modflowapi](https://github.com/MODFLOW-USGS/modflowapi) and [phreeqcrm](https://github.com/usgs-coupled/phreeqcrm)
