### 2026-03-18 #9762 @p-ortega
Hi @kthyng

- Testing is inside [here](https://github.com/p-ortega/mf6rtm/tree/JOSS/autotest)
- Those are actually dependencies being installed with mf6rtm. They are not included in the codebase.
