+++ JOSS/paper.bib
@@ -5,7 +5,7 @@ volume = {109},
 pages = {191-201},
 year = {2018},
 issn = {1364-8152},
-doi = {https://doi.org/10.1016/j.envsoft.2018.06.009},
+doi = {10.1016/j.envsoft.2018.06.009},
 url = {https://www.sciencedirect.com/science/article/pii/S1364815218302676},
 author = {Jeremy T. White},
 keywords = {Modeling, Uncertainty, Iterative ensemble smoother, Gauss-Levenberg-Marquardt},
