### 2026-03-25 #9762 @p-ortega
@kthyng  I apologise for that. Those local deps were legacy from early development efforts. I have removed them now and updated the JOSS branch.
