### 2026-08-06 259e5d6
fix: hdf5 refactor broke emulator vars outputs, moved downstream
