### 2026-08-17 #9762 @martinvonk
Thought I’d chime in here. My experience is mainly with MODFLOW and the MODFLOW 6 API, though I have (very) basic PHREEQC experience from university. Anyone with a solid chemistry background would be a great fit for the second reviewer, which might open up some more options for you @mengqi-z.
