### 2026-08-17 #9762 @mengqi-z
@p-ortega @martinvonk - Thank you both! And my apologies for taking a while to find the second reviewer. I have been reaching out offline as well to potential reviewers with MODFLOW and reactive transport experience, but haven't had much success unfortunately.

Both potential reviewers suggested by @p-ortega seem like great fits. I'll also expand my next round of search to include people with chemistry background, as @martinvonk suggested. Thank you!
