### 2026-08-17 #9762 @mengqi-z
👋 @mwtoews @vcantarella - Would any of you be interested and available to review this submission titled "MF6RTM: a python package for predictive reactive transport modeling via the MODFLOW 6 and PHREEQC APIs" for the Journal of Open Source Software (JOSS)?
