### 2026-08-19 #9762 @vcantarella
Hi. I will gladly review the article. let me know if I can proceed.

Cheers,

Vitor C.,
