+++ mf6rtm/simulation/solver.py
@@ -16,6 +16,8 @@ from mf6rtm.simulation.discretization import total_cells_in_grid
 from mf6rtm.config.config import MF6RTMConfig
 from mf6rtm.io.externalio import SelectedOutput
 from mf6rtm.utils import utils
+from mf6rtm.simulation.ddmt import validate_ddmt_arrays, exchange_first_order_single_rate
+from mf6rtm.io.ddmt_output import DDMTOutputWriter                                                                                         
 
 # warnings.filterwarnings("ignore")
 # warnings.filterwarnings("ignore", category=DeprecationWarning)
@@ -348,6 +350,25 @@ class Mf6RTM(object):
         self.set_time_conversion()
 
         self.config = MF6RTMConfig.from_toml_file(self.wd/"mf6rtm.toml")
+
+        ### ddmt
+        ddmt_cfg = getattr(self.config, "ddmt", {})
+
+        self.ddmt_enabled = ddmt_cfg.get("enabled", False)
+        self.ddmt_mode = ddmt_cfg.get("mode", "storage")
+        self.theta_mobile = ddmt_cfg.get("theta_mobile", None)
+        self.theta_immobile = ddmt_cfg.get("theta_immobile", None)
+        self.alpham = ddmt_cfg.get("alpham", None)
+
+        self.immobile_initial = ddmt_cfg.get("immobile_initial", "mobile")
+        self.immobile_conc_m3 = None
+
+        self.immobile_yaml = ddmt_cfg.get("immobile_yaml", None)
+        self.phreeqcbmi_immobile = None
+
+        self.ddmt_output_enabled = ddmt_cfg.get("output", True)
+        self.ddmt_output_format = ddmt_cfg.get("output_format", "hdf5")
+        self.ddmt_output = None
         self.reactive = self.config.reactive['enabled']
         # diffmask threshold: cells changing less than this skip reactions
         self.threshold = self.config.solver.get('threshold', self.threshold)
@@ -525,6 +546,80 @@ class Mf6RTM(object):
         """Set the model to run only transport or transport and reactions"""
         self.reactive = reactive
 
+
+    ### ddmt
+    def _prepare_ddmt(self):
+        """Validate DDMT settings and prepare optional immobile chemistry/output."""
+        if not self.ddmt_enabled:
+            return
+
+        self.theta_mobile, self.theta_immobile, self.alpham = validate_ddmt_arrays(
+            self.theta_mobile,
+            self.theta_immobile,
+            self.alpham,
+            self.nxyz,
+        )
+
+        if self.ddmt_mode not in ("storage", "reactive"):
+            raise ValueError("ddmt_mode must be either 'storage' or 'reactive'")
+
+        if self.ddmt_mode == "reactive":
+            self._prepare_immobile_phreeqcrm()
+
+        if self.ddmt_output_enabled:
+            self.ddmt_output = DDMTOutputWriter(
+                self,
+                output_format=self.ddmt_output_format,
+            )
+
+
+    def _prepare_immobile_phreeqcrm(self):
+        """Prepare the second PhreeqcBMI instance for immobile-domain reactions."""
+        if self.phreeqcbmi_immobile is None:
+            if self.immobile_yaml is None:
+                raise ValueError(
+                    "ddmt_mode='reactive' requires immobile_yaml or phreeqcbmi_immobile"
+                )
+            self.phreeqcbmi_immobile = PhreeqcBMI(self.immobile_yaml)
+
+        self.phreeqcbmi_immobile._prepare_phreeqcrm_bmi()
+        self.phreeqcbmi_immobile.SetTimeConversion(self.time_conversion)
+        
+        mobile_components = list(self.phreeqcbmi.components)
+        immobile_components = list(self.phreeqcbmi_immobile.components)
+
[…]
