### 2026-08-24 #11173 @vcantarella
Thank you @martinvonk. I will start taking a look later today. I will probably need the 4 weeks to give a proper feedback. @mengqi-z I noticed the packaged development has continued and the JOSS branch is way behind, about 120 commits. should I give feedback to that branch, nevertheless?
