### 2026-08-24 #11173 @mengqi-z
@vcantarella - Thanks for the question. The `JOSS` branch is just the branch that holds the paper file `paper.md`. You should review the `main` branch where the main development occurs.
