### 2026-08-24 e6a269d
packaging: consolidate build/dependency config into pyproject.toml
Replace setup.cfg/setup.py/pixi.toml/.bumpversion.toml with a single
pyproject.toml (hatchling + hatch-vcs, PyPI deps mirrored to
tool.pixi.dependencies), migrate lint from flake8 to ruff, and fold
autotest/pytest.ini into [tool.pytest.ini_options]. Every pixi
environment now editable-installs mf6rtm automatically, fixing the
"pixi install doesn't make mf6rtm importable" part of #81. Also
narrows supported Python to >=3.11,<3.14 and updates CI matrices to
match, and adds authors/maintainer metadata.

Apply `ruff check --fix` ac
[…]
