+++ mf6rtm/io/externalio.py
@@ -3,12 +3,14 @@ inputs from layered txt files.
 """
 import os
 import re
+
 import numpy as np
 import pandas as pd
-from mf6rtm.simulation.mf6api import Mf6API
-from mf6rtm.simulation.discretization import grid_dimensions, total_cells_in_grid
-from mf6rtm.config.yaml_reader import load_yaml_to_phreeqcrm
+
 from mf6rtm.config.config import MF6RTMConfig
+from mf6rtm.config.yaml_reader import load_yaml_to_phreeqcrm
+from mf6rtm.simulation.discretization import grid_dimensions, total_cells_in_grid
+from mf6rtm.simulation.mf6api import Mf6API
 from mf6rtm.utils.utils import get_indices
 
 ic_position = {
@@ -352,8 +354,8 @@ class Regenerator:
         for i_phase in range(1, n_phases+1):
             block = f"EQUILIBRIUM_PHASES {i_phase}\n"
             for nme in equilibrium_phases['names']:
-                si = equilibrium_phases.get(f'si', None).get(nme, None)
-                m0 = equilibrium_phases.get(f'm0', None).get(nme, None).flatten()
+                si = equilibrium_phases.get('si', None).get(nme, None)
+                m0 = equilibrium_phases.get('m0', None).get(nme, None).flatten()
                 block += f"    {nme} {si:.5e} {m0[i_phase-1]:.5e}\n"
             block += "END\n"
             blocks.append(block)
