+++ mf6rtm/simulation/discretization.py
@@ -2,8 +2,9 @@
 This module contains the Discretization functions for various ModFlow grid types.
 """
 
-from mf6rtm.simulation.mf6api import Mf6API
 import math
+
+from mf6rtm.simulation.mf6api import Mf6API
 
 
 def total_cells_in_grid(modflow_api: Mf6API) -> int:
