+++ mf6rtm/assets/__init__.py
@@ -1,5 +1,6 @@
 from importlib import resources
 
+
 def mrbeaker_path():
     with resources.path("mf6rtm.assets", "mrbeaker.png") as path:
         return path
