### 2026-08-24 a775db9
ci: bump pinned pixi CLI version to match the lock file
pixi.lock is now schema version 7 (written by a current pixi), but
every workflow pinned pixi-version: v0.41.4, which only supports lock
file versions up to 6. Every CI job was failing in "Setup pixi" before
any tests ran. Bump the pin to v0.76.0 in ci-pixi.yml, ci-pixi-macos.yml,
and docs.yml.
