+++ .github/workflows/ci-pixi-macos.yml
@@ -24,7 +24,7 @@ jobs:
       uses: prefix-dev/setup-pixi@v0.9.3
       with:
         log-level: vvv
-        pixi-version: v0.41.4
+        pixi-version: v0.76.0
         cache: true
         environments: py${{ matrix.python-version }}
 
