+++ .github/workflows/ci-pixi.yml
@@ -23,7 +23,7 @@ jobs:
     - name: Setup pixi
       uses: prefix-dev/setup-pixi@v0.9.3
       with:
-        pixi-version: v0.41.4
+        pixi-version: v0.76.0
         cache: true
         environments: py${{ matrix.python-version }}
 
@@ -40,7 +40,7 @@ jobs:
     - name: Setup pixi
       uses: prefix-dev/setup-pixi@v0.9.3
       with:
-        pixi-version: v0.41.4
+        pixi-version: v0.76.0
         cache: true
 
     - name: Run linting
@@ -54,7 +54,7 @@ jobs:
     - name: Setup pixi
       uses: prefix-dev/setup-pixi@v0.9.3
       with:
-        pixi-version: v0.41.4
+        pixi-version: v0.76.0
         cache: true
         environments: py311
 
