### 2026-08-24 34f6a72
deps: fix docs pixi feature gaps, add pyarrow to core deps
Add jupytext/nbconvert/ipykernel (needed by exec-tutorials' jupytext
--execute step), shapely (an optional flopy dep one tutorial's
plotting code triggers), and pytables (one tutorial exercises
mf6rtm's hdf5 output path) to the docs pixi feature -- all were only
available via the old flat pixi.toml's single dependency list before
this packaging consolidation split things into dev/docs features.
Verified by running `pixi run -e docs build-docs` locally until it
completed clean.

Also add pyarrow to [project.dependencies]/[tool.pixi.depe
[…]
