### 2026-08-24 13ef854
build: prep a conda-forge recipe (not submitted)
Add conda-recipe/meta.yaml as an unsubmitted starting point for a
future conda-forge/staged-recipes PR, plus a `pixi run
build-conda-recipe` task (grayskull against a local sdist) to
regenerate it. Add python-build/grayskull as real dev dependencies.

Fixed by hand after generation, and documented for future
regenerations: grayskull points `source` at a local file:// sdist
(no PyPI release with the corrected dependency list exists yet) --
repointed at the latest real release (0.5.2) instead. It also adds a
`mf6rtm --help` test command that doesn
[…]
