### 2026-08-24 4d2c183
docs: clean up install instructions (closes #81)
Remove benchmark/readme.md's "Alternate Approach using only Conda"
section and benchmark/environment.yml -- a second, Anaconda-based
install path that conflicted with the pixi-first workflow and is
what #81 flagged as confusing. Update README.md's Developing section
to reflect that `pixi install` alone now gives a working, importable
environment, and note that pixi run test/test-cov fetch MODFLOW
binaries automatically. Add a short "from conda-forge" note marked
as not yet published.
