### 2026-08-24 690eb59
build: add jupytext pairing config, trim pyproject.toml comments
Add [tool.jupytext] (scoped to docs/tutorials/ only) -- verified that
docs/tutorials/*.py <-> *.ipynb had no pairing declared anywhere, so
neither `jupytext --sync` nor an editor's Jupytext-Sync extension
could find it; it only worked via exec-tutorials' one-way
`--to notebook --execute`. Move the fuller rationale for pyproject.toml's
various packaging/pixi decisions into AGENTS.md, trimming inline
comments to one-line pointers. Fix two inaccuracies found while
reviewing AGENTS.md (benchmark/ has no paired .py; autotest/ doesn't

[…]
