### 2026-08-26 4112764
ci: add Python 3.14 support, add AGENTS.md guardrails section
Extends requires-python/classifiers, the py314 pixi feature+environment,
and both CI matrices to cover 3.14. Fixes benchmark/ex7.ipynb's use of
matplotlib.cm.get_cmap, removed in matplotlib 3.11 (which the fresh
py314 solve pulls in), so the new CI job actually passes.

Also adds a Critical guardrails section to AGENTS.md covering
commit/merge/release ownership and multi-step review pauses, mirroring
the sibling data-engine repo's convention.
