### 2026-08-31 2856fe2
Add `min_concentration` to tutorials
Begining to address #82, but problem persists because this is not caused by numerical diffussion from transport model but rather by depleting carbonate buffering capacity.
