### 2026-09-07 73e5930
Add CITATION.cff
