### 2026-09-18 9ab4f54
minor check fixes
