+++ mf6rtm/config/config.py
@@ -101,13 +101,13 @@ class MF6RTMConfig:
                 "enabled": False,
                 "mode": "storage",
                 "theta_mobile": None,
-                "theta_immobile": None, 
+                "theta_immobile": None,
                 "alpham": None,
                 "immobile_initial": "mobile",
                 "immobile_yaml": None,
                 "output": True,
                 "output_format": "hdf5",
-            },                     
+            },
         }
 
         for section, section_defaults in defaults.items():
@@ -421,7 +421,7 @@ class MF6RTMConfig:
                 "immobile_yaml": ddmt_config.get("immobile_yaml", None),
                 "output": ddmt_config.get("output", True),
                 "output_format": ddmt_config.get("output_format", "hdf5"),
-            }                                                             
+            }
 
         # Flatten everything *except* known sections
         remaining_dict = {k: v for k, v in config_dict.items() if k not in ['reactive', 'solver', 'output',
